:hammer: Refactor ui code to package by interface
