+++ composeApp/src/commonMain/kotlin/com/clipevery/config/AppConfig.kt
@@ -7,7 +7,6 @@ import java.util.UUID
 @Serializable
 data class AppConfig(
     val appInstanceId: String = UUID.randomUUID().toString(),
-    val bindingState: Boolean = false,
     val language: String = Locale.getDefault().language,
     val isFollowSystemTheme: Boolean = true,
     val isDarkTheme: Boolean = false
